+++ packages/mutation-testing-metrics-scala/metrics/src/main/scala/mutationtesting/MutantStatus.scala
@@ -1,0 +1,7 @@
+package mutationtesting
+
+object MutantStatus extends Enumeration {
+  type MutantStatus = Value
+
+  val Killed, Survived, NoCoverage, Timeout, CompileError = Value
+}